+++ CHANGELOG.rst
@@ -134,6 +134,9 @@ Added
   `#266 <https://github.com/openego/eGon-data/issues/266>`_
 * Add random seed to CLI parameters
   `#351 <https://github.com/openego/eGon-data/issues/351>`_
+* Extend zensus by a combined table with all cells where
+  there's either building, apartment or population data
+  `#359 <https://github.com/openego/eGon-data/issues/359>`_
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159